Updated pathfinding
Updated pathfinding - removed a lot of seemingly excess logic. There
were 2 strategies for pathfinding, one for finding paths that were only
a single cell away, and one for finding paths to further cells.
Consolidated this into a single path.

Updated the logic to decide whether or not it should go to a certain
cell. Now if it does not have enough energy to go to a cell and vacuum
it, the vacuum will not go to that cell.

Lowered the power to 50 so it is now unable to clean the entire floor.